% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_missing_phenotypes.R
\name{add_missing_phenotypes}
\alias{add_missing_phenotypes}
\title{Add missing phenotypes to samples}
\usage{
add_missing_phenotypes(
  mt,
  all_samples_info,
  skip_descr = FALSE,
  skip_mat = FALSE
)
}
\arguments{
\item{mt}{A data frame containing the samples.}

\item{skip_descr}{Logical indicating whether to skip adding
Sample.Description (default is FALSE).}

\item{skip_mat}{Logical indicating whether to skip adding
Material (default is FALSE).}
}
\value{
The data frame with missing phenotypes added.
}
\description{
This function adds missing phenotypes to samples based on information
 from a metadata table.
}
