% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_by_key}
\alias{aggregate_by_key}
\title{Automatic data_set aggregation by key}
\usage{
aggregate_by_key(data_set, key, verbose = TRUE, thresh = 53, ...)
}
\arguments{
\item{data_set}{Matrix, data.frame or data.table (with only numeric, integer, factor, logical, character columns)}

\item{key}{Name of a column of data_set according to which the set should be aggregated (character)}

\item{verbose}{Should the algorithm talk? (logical, default to TRUE)}

\item{thresh}{Number of max values for frequencies count (numerical, default to 53)}

\item{...}{Optional argument: \code{functions}:  aggregation functions for numeric columns
(vector of function names (character), optional, if not set we use: c("mean", "min", "max", "sd"))}
}
\value{
A \code{\link[data.table]{data.table}} with one line per \code{key} elements and multiple  new columns.
}
\description{
Automatic aggregation of a data_set set according to a \code{key}.
}
\details{
Perform aggregation depending on column type:
\itemize{
  \item If column is numeric \code{functions} are performed on the column. So 1 numeric column
    give length(functions) new columns,
  \item If column is character or factor and have less than \code{thresh} different values,
    frequency count of values is performed,
  \item If column is character or factor with more than \code{thresh} different values, number
    of different values for each \code{key} is performed,
  \item If column is logical, number of TRUE is computed.
}

In all cases, if the set as more rows than unique \code{key}, a number of lines will be computed.

Be careful using functions argument, given functions should be an aggregation function,
meaning that for multiple values it should only return one value.
}
\examples{
\dontrun{
# Get generic dataset from R
data("adult")

# Aggregate it using aggregate_by_key, in order to extract characteristics for each country
adult_aggregated <- aggregate_by_key(adult, key = 'country')

# Example with other functions
power <- function(x) {sum(x^2)}
adult_aggregated <- aggregate_by_key(adult, key = 'country', functions = c("power", "sqrt"))

# sqrt is not an aggregation function, so it wasn't used.
}
# "##NOT RUN:" mean that this example hasn't been run on CRAN since its long. But you can run it!
}
