% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{apply_bins}
\alias{apply_bins}
\title{Apply discretization}
\usage{
apply_bins(dataSet, bins, verbose = TRUE)
}
\arguments{
\item{dataSet}{Matrix, data.frame or data.table}

\item{bins}{vector of bins for discretization. They should come from \code{\link{get_bins}} \cr
or be carefully hand written.}

\item{verbose}{Should the algorithm talk? (Logical, default to TRUE)}
}
\value{
dataSet with numerical columns discretized by \strong{reference}.
}
\description{
Apply discretization using pre-set bins.
}
\details{
For now, opening and closing brackets are used the same way.
}
\examples{
# Build discretized data
data(adult)
adult <- fastDiscretization(adult)
bins <- get_bins(adult)

# Apply it to messy_adult
data(messy_adult)
messy_adult <- apply_bins(messy_adult, bins = bins)

# To apply your own bins on column "age"
data(messy_adult)
messy_adult <- apply_bins(messy_adult, bins = list(age = c("[0,40[","[40, 100]")))
}
