% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateFromFactor.R
\name{build_encoding}
\alias{build_encoding}
\title{Compute encoding}
\usage{
build_encoding(dataSet, cols = "auto", verbose = TRUE, ...)
}
\arguments{
\item{dataSet}{Matrix, data.frame or data.table}

\item{cols}{List of numeric column(s) name(s) of dataSet to transform. To transform all 
characters, set it to "auto". (character, default to "auto")}

\item{verbose}{Should the algorithm talk? (Logical, default to TRUE)}

\item{...}{Other arguments such as \code{name_separator} to separate words in new columns names
(character, default to ".")}
}
\value{
A list where each element name is a column name of data set and each element new_cols 
and values the new columns that will be built during encoding.
}
\description{
Build a list of one hot encoding for each \code{cols}.
}
\examples{
# Get a data set
data(adult)
encoding <- build_encoding(adult, cols = "auto", verbose = TRUE)

print(encoding)
}
