% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_dataset_inspection.R
\name{check_memory_usage}
\alias{check_memory_usage}
\title{Check Memory Usage of Datasets}
\usage{
check_memory_usage(envir = .GlobalEnv)
}
\arguments{
\item{envir}{Environment to search (default: global environment)}
}
\value{
Invisible data frame with dataset information
}
\description{
Lists all data frames in the global environment and their memory usage.
Useful for identifying large datasets.
}
\examples{
\dontrun{
df1 <- data.frame(x = 1:1000)
df2 <- data.frame(y = 1:10000)
check_memory_usage()
}

}
