% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corrected_list}
\alias{corrected_list}
\title{Batch cerration list for Package Examples}
\format{
A list with 2 datasets:
  \describe{
    \item{Before correction}{A dataset containing raw data before
    batch correction}
    \item{removeBatchEffect}{A dataset after correcting raw data using
    removeBatchEffect function from `limma` R package.}
    }
}
\usage{
corrected_list
}
\description{
This list contains two datasets used for demonstrating examples
 in the package.
}
\keyword{datasets}
