% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_numeric_features.R
\name{create_quantile_features}
\alias{create_quantile_features}
\title{Create Quantile Features}
\usage{
create_quantile_features(data, col)
}
\arguments{
\item{data}{A data frame}

\item{col}{Column name (character)}
}
\value{
Data frame with quantile column
}
\description{
Create Quantile Features
}
