% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_numeric_features.R
\name{create_rank_features}
\alias{create_rank_features}
\title{Create Rank Features}
\usage{
create_rank_features(data, col, group_col = NULL)
}
\arguments{
\item{data}{A data frame}

\item{col}{Column name (character)}

\item{group_col}{Optional grouping column}
}
\value{
Data frame with rank column
}
\description{
Create Rank Features
}
