% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_logger.R
\name{DataPreparation_getLogger}
\alias{DataPreparation_getLogger}
\title{Retrieves DataPreparation logger.}
\usage{
DataPreparation_getLogger()
}
\value{
logger object
}
\description{
Retrieves DataPreparation logger.
}
