% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description.R
\name{description}
\alias{description}
\title{Describe data set}
\usage{
description(dataSet, level = 1, path_to_write = NULL, verbose = TRUE)
}
\arguments{
\item{dataSet}{Matrix, data.frame or data.table}

\item{level}{Level of description (0: generic, 1: column by column) 
(numeric, default to 1)}

\item{path_to_write}{Path where the report should be written (character, default to NULL)}

\item{verbose}{Should the algorithm talk? (Logical, default to TRUE)}
}
\description{
Generate extensive description of a data set.
}
\examples{
# Load exemple set
data(messy_adult)

# Describe it
description(messy_adult)
}
