% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_numeric_features.R
\name{detect_extract_half_full}
\alias{detect_extract_half_full}
\title{Detect Half vs Full Ratings}
\usage{
detect_extract_half_full(data, col)
}
\arguments{
\item{data}{A data frame}

\item{col}{Column name (character)}
}
\value{
Data frame with half/full flag
}
\description{
Detect Half vs Full Ratings
}
