% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastFunctions.r
\name{fastHandleNa}
\alias{fastHandleNa}
\title{Handle NA values}
\usage{
fastHandleNa(dataSet, codeToSetNumeric = 0, codeToSetBooleans = FALSE,
  codeToSetCharacter = "\\"\\"", verbose = TRUE)
}
\arguments{
\item{dataSet}{Matrix, data.frame or data.table}

\item{codeToSetNumeric}{NAs replacement in numeric column, (numeric or code in a string, default to 0)}

\item{codeToSetBooleans}{NAs replacement in logical column, (logical or code in a string, default to FALSE)}

\item{codeToSetCharacter}{NAs replacement in character column, (character or code in a string, default to '""')}

\item{verbose}{Should the algorithm talk (logical, default to TRUE)}
}
\value{
dataSet as a \code{\link{data.table}} with NAs handled
}
\description{
Function to handle NAs values depending on the class of the column
}
\details{
To preserve RAM this function edit directly the dataSet set.  
For factor columns, it will add NA to list of values.
}
\examples{
# Build a useful dataSet set for example
dataSet <-  data.table(numCol = c(1, 2, 3, NA),
                   charCol = c("", "a", NA, "c"),
                   booleanCol = c(TRUE, NA, FALSE, NA))

# To set NAs to 0, FALSE and "" (respectively for numeric, logical, character)
dataSet <- fastHandleNa(dataSet)

# In a numeric column to set NAs as "missing"
dataSet <-  data.table(numCol = c(1, 2, 3, NA),
                   charCol = c("", "a", NA, "c"),
                   booleanCol = c(TRUE, NA, FALSE, NA))
dataSet <- fastHandleNa(dataSet, codeToSetCharacter = '"missing"')

# In a numeric column, to set NAs to the minimum value of the column
dataSet <-  data.table(numCol = c(1, 2, 3, NA),
                   charCol = c("", "a", NA, "c"),
                   booleanCol = c(TRUE, NA, FALSE, NA))
dataSet <- fastHandleNa(dataSet, codeToSetNumeric = "min(dataSet[[col]], na.rm = TRUE)")

}
