% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{fastUnScale}
\alias{fastUnScale}
\title{unScale}
\usage{
fastUnScale(dataSet, scales, verbose = TRUE)
}
\arguments{
\item{dataSet}{Matrix, data.frame or data.table, output of \code{\link{fastScale}}}

\item{scales}{Result of funcion \code{\link{build_scales}}. \cr
To perform the same unscaling on train and test, it is compulsory to compute \code{\link{build_scales}}
before.}

\item{verbose}{Should the algorithm talk? (Logical, default to TRUE)}
}
\value{
\code{dataSet} with columns unScaled by \strong{reference}, this will make the overall 
values back to actual values, that were before the scale.
}
\description{
Perform efficient unScaling of the scaled data set which will be output of \code{\link{fastScale}}.
}
\details{
unScaling numeric values is very usefull for most post-model analysis
}
\examples{
# Load data
data(adult)

# compute scales
scales <- build_scales(adult, cols = "auto", verbose = TRUE)

originalAgeMean = mean(adult$age) 
originalAgeSD   = sd(adult$age)

# Scale data set
adult <- fastScale(adult, scales = scales, verbose = TRUE)

# Control
print(mean(adult$age)) # Almost 0
print(sd(adult$age)) # 1

adult <- fastUnScale(adult, scales = scales, verbose = TRUE)

mean(adult$age) == originalAgeMean # TRUE
sd(adult$age)   == originalAgeSD   # TRUE
}
