% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_na_handling.R
\name{flag_na_column}
\alias{flag_na_column}
\title{Flag NA Values in Column}
\usage{
flag_na_column(data = NULL, col, dataset_name = NULL)
}
\arguments{
\item{data}{A data frame (optional if dataset_name provided)}

\item{col}{Column name (character)}

\item{dataset_name}{Character. Name of registered dataset to use.}
}
\value{
Data frame with new NA flag column
}
\description{
Flag NA Values in Column
}
