% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncols.R
\name{fncols}
\alias{fncols}
\title{Fill Missing Columns with NA Values}
\usage{
fncols(data, cname)
}
\arguments{
\item{data}{A data frame.}

\item{cname}{A character vector specifying column names to be filled with NA values.}
}
\value{
A data frame with missing columns filled with NA values.
}
\description{
This function fills missing columns in a data frame with NA values based on the specified column names.
}
