% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_dataset_registry.R
\name{focus_dataset}
\alias{focus_dataset}
\title{Set Focus on a Dataset}
\usage{
focus_dataset(dataset_name, verbose = TRUE)
}
\arguments{
\item{dataset_name}{Character string. Name of the registered dataset to focus on.
Must be a registered dataset (use `register_dataset()` first).}

\item{verbose}{Logical. If TRUE, prints confirmation message. Default: TRUE}
}
\value{
Invisibly returns the registry
}
\description{
Sets the "current" dataset that all subsequent functions will use by default.
Functions will target this focused dataset unless another dataset is explicitly
passed via parameters.
}
\details{
Use this to streamline workflows where you're working primarily with one dataset.
When done, call `stop_focus()` to disable focus mode.


**Focus Mode Workflow:**
\preformatted{
register_dataset(df1, "movies")
register_dataset(df2, "reviews")

# Set focus on movies
focus_dataset("movies")

# These functions now use 'movies' by default:
inspect_dataset()                # Uses movies (when updated in Phase 2)
generate_ngrams(string_cols = "title")  # Uses movies

# Switch focus to reviews
focus_dataset("reviews")

# Now uses reviews instead:
inspect_dataset()  # Now uses reviews

# Disable focus mode
stop_focus()

# Back to regular mode - must pass data explicitly
inspect_dataset(data = df1)
}
}
\examples{
\dontrun{
movies <- data.frame(title = c("Matrix", "Inception"))
cancer <- data.frame(diagnosis = c("Breast", "Lung"))

register_dataset(movies, "movies")
register_dataset(cancer, "cancer")

# Set focus on movies
focus_dataset("movies")
get_current_focus()  # "movies"

# Switch to cancer
focus_dataset("cancer")
get_current_focus()  # "cancer"

# Disable focus
stop_focus()
get_current_focus()  # NULL
}

}
