% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_numeric_features.R, R/04_string_features.R
\name{generate_categorical_variants}
\alias{generate_categorical_variants}
\title{Generate Categorical Encoding Variants}
\usage{
generate_categorical_variants(data, label_cols)

generate_categorical_variants(data, label_cols)
}
\arguments{
\item{data}{A data frame containing categorical columns.}

\item{label_cols}{Character vector of categorical column names to encode.}

\item{dataset_name}{Character. Name of registered dataset to use.}
}
\value{
List of encoding variant datasets

A named list of candidate datasets with different categorical encodings.
}
\description{
Create candidate datasets with categorical encoding methods such as
one-hot encoding for categorical columns.
}
\examples{
\dontrun{
candidates <- generate_categorical_variants(iris, "Species")
}

}
