% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_numeric_features.R
\name{generate_numeric_variants_advanced}
\alias{generate_numeric_variants_advanced}
\title{Generate Advanced Numeric Variants}
\usage{
generate_numeric_variants_advanced(
  data = NULL,
  numeric_cols,
  include_poly = TRUE,
  include_transforms = TRUE,
  include_bins = TRUE,
  include_outliers = TRUE,
  dataset_name = NULL
)
}
\arguments{
\item{data}{A data frame (optional if dataset_name provided)}

\item{numeric_cols}{Vector of numeric column names}

\item{include_poly}{Include polynomial features}

\item{include_transforms}{Include transformations}

\item{include_bins}{Include binning features}

\item{include_outliers}{Include outlier flags}

\item{dataset_name}{Character. Name of registered dataset to use.}
}
\value{
List of advanced variant datasets
}
\description{
Generate Advanced Numeric Variants
}
