% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_workflow_generator.R
\name{generate_workflow_script}
\alias{generate_workflow_script}
\title{Generate Workflow Script Code}
\usage{
generate_workflow_script(
  dataset_name,
  string_col,
  numeric_col,
  id_col,
  analysis_type,
  data
)
}
\arguments{
\item{dataset_name}{Name of the dataset}

\item{string_col}{Name of string column for text analysis}

\item{numeric_col}{Name of numeric column for target variable}

\item{id_col}{Name of ID column}

\item{analysis_type}{Type of analysis to perform}

\item{data}{The actual dataset (for column validation)}
}
\value{
Character string containing complete R script
}
\description{
Creates R code for a complete workflow based on selected parameters.
}
\keyword{internal}
