% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateFromDate.R
\name{generateFactorFromDate}
\alias{generateFactorFromDate}
\title{Generate factor from dates}
\usage{
generateFactorFromDate(dataSet, cols, type = "yearmonth", verbose = TRUE,
  ...)
}
\arguments{
\item{dataSet}{Matrix, data.frame or data.table}

\item{cols}{a list of colnames of dataSet (or just one) to transform into numerics}

\item{type}{"year", "yearquarter", "yearmonth", "quarter" or "month", way to aggregate a date, 
(character, default to "yearmonth")}

\item{verbose}{should the function log (logical, default to TRUE)}
}
\description{
Taking Date or POSIXct colums, and building factor columns from them.
}
\examples{
# Load set, and find dates
data(messy_adult)
messy_adult <- findAndTransformDates(messy_adult, verbose = FALSE)

# Generate new columns
# Genereate year month columns
messy_adult <- generateFactorFromDate(messy_adult, cols = c("date1", "date2", "num1"))
head(messy_adult[, .(date1.yearmonth, date2.yearmonth)])


# Genereate quarter columns
messy_adult <- generateFactorFromDate(messy_adult, cols = c("date1", "date2"), type = "quarter")
head(messy_adult[, .(date1.quarter, date2.quarter)])
}
