% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{get_bins}
\alias{get_bins}
\title{Get bins}
\usage{
get_bins(dataSet)
}
\arguments{
\item{dataSet}{Matrix, data.frame or data.table}
}
\value{
List of list where each element name is the column name and the element is a the list of bins. \cr
}
\description{
Retriveve bins from a data set containing discretized columns (using function \code{\link{fastDiscretization}})
}
\examples{
# Build a discretized data set
data(adult)
adult <- fastDiscretization(adult, verbose = FALSE)

# Retrieve bins
get_bins(adult)
}
