% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_manipulation.R
\name{get_most_frequent_element}
\alias{get_most_frequent_element}
\title{Get most frequent element}
\usage{
get_most_frequent_element(x)
}
\arguments{
\item{x}{A list, \code{data.frame} or \code{data.table} column}
}
\value{
The most frequent element
}
\description{
Provide most frequent element in a list, a \code{data.frame} or \code{data.table} 
column
}
\examples{
# Build example list
example_list <- c(1, 1, 2, 3, 1, 4, 1)

# Compute most frequent element
get_most_frequent_element(example_list)
}
