% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_dataset_registry.R
\name{get_registered_dataset}
\alias{get_registered_dataset}
\title{Get Registered Dataset}
\usage{
get_registered_dataset(name, return_data = TRUE)
}
\arguments{
\item{name}{Character string. Name of the dataset to retrieve}

\item{return_data}{Logical. If TRUE, returns just the data frame.
If FALSE, returns full metadata. Default: TRUE}
}
\value{
Data frame (if return_data=TRUE) or metadata list
}
\description{
Retrieves a dataset from the registry by name.
Use this internally in functions to access registered data.
}
\examples{
\dontrun{
# Get just the data
movies_data <- get_registered_dataset("movies")

# Get full metadata
movies_metadata <- get_registered_dataset("movies", return_data = FALSE)
}

}
\keyword{internal}
