% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_values.R
\name{impute_missing_values}
\alias{impute_missing_values}
\title{Impute missing values using k-nearest neighbors algorithm}
\usage{
impute_missing_values(data, k = 5)
}
\arguments{
\item{data}{The input data frame.}

\item{k}{The number of nearest neighbors to consider (default is 5).}
}
\value{
The data frame with missing values imputed.
}
\description{
This function imputes missing values in the dataset using
the k-nearest neighbors algorithm.
}
