% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_na_handling.R
\name{impute_na_numeric}
\alias{impute_na_numeric}
\title{Impute NA in Numeric Column with Median}
\usage{
impute_na_numeric(data, col)
}
\arguments{
\item{data}{A data frame}

\item{col}{Column name (character)}
}
\value{
Data frame with NA values imputed with median
}
\description{
Impute NA in Numeric Column with Median
}
\examples{
\dontrun{
impute_na_numeric(mtcars, "mpg")
}

}
