% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imputed_data}
\alias{imputed_data}
\title{The data frame with missing values imputed}
\format{
A data frame with `m` observations(in rows) and
 `n` variables(in columns):
  \describe{
    \item{...}{More numeric variables (Arg, Asn, Asp, Cit, Gln, Glu, Gly,
         His, Ile, Lys, Met, Orn, Phe, Pro, Ser, Thr, Trp, Tyr, Val, xLeu,
         Ac.Orn, ADMA, alpha.AAA, c4.OH.Pro, Carnosine, Creatinine, DOPA,
         Dopamine, Histamine, Kynurenine, Met.SO, Nitro.Tyr, PEA, Putrescine,
         Sarcosine, SDMA, Serotonin, Spermidine, Spermine, t4.OH.Pro,
         Taurine, CE, DG, TG, LPC, PC, Cer, SM, H1).}
  }
}
\usage{
imputed_data
}
\description{
This dataset contains impute data missing values using k-nearest
neighbors algorithm.
}
\keyword{datasets}
