% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSet.R
\docType{data}
\name{messy_adult}
\alias{messy_adult}
\title{Adult with some ugly columns added}
\format{A data.table with 32561 rows and 24 variables.}
\usage{
data("messy_adult")
}
\description{
For examples and tutorials, messy_adult has been built using UCI \code{adult}.\cr
}
\details{
We added 9 really ugly columns to the data set:

\itemize{
    \item 4 dates with various formats and time stamp, containing NAs
    \item 1 constant column
    \item 3 numeric with different decimal separator
    \item 1 email address
}
}
\keyword{data}
