% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_pca.R
\name{perform_pca}
\alias{perform_pca}
\title{Perform Principal Component Analysis (PCA)}
\usage{
perform_pca(data, scale_data = TRUE, center_data = TRUE)
}
\arguments{
\item{data}{The input data frame.}

\item{scale_data}{Logical indicating whether to scale the data
(default is TRUE).}

\item{center_data}{Logical indicating whether to center the data
(default is TRUE).}
}
\value{
A list containing the PCA results, including the principal
component scores and other attributes.
}
\description{
This function performs Principal Component Analysis (PCA) on the input data.
}
