% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Process Data}
\usage{
process_data(data, categories)
}
\arguments{
\item{data}{A matrix or data frame containing the input data.}

\item{categories}{A data frame containing categorical information for the samples.}
}
\value{
A list containing processed data and corresponding PCA plots.
}
\description{
Process the input data using various methods including filtering, CLR transformation, batch effect correction, and PCA.
}
\examples{
# Generate sample data
set.seed(123)
data <- data.frame(
  Sample.Type = sample(c("Type1", "Type2", "Type3"), 100, replace = TRUE),
  Sample.Identification = paste0("ID", 1:100),
  Gender = sample(c("Female", "Male"), 100, replace = TRUE),
  Age = sample(20:80, 100, replace = TRUE),
  Sample.Description = sample(c("DLB", "msAD", "miAD", "HC"), 100, replace = TRUE),
  Submission.Name = sample(c("plate1", "plate2"), 100, replace = TRUE),
  Material = sample(c("serum", "sample"), 100, replace = TRUE),
  Apoe.Genotype = sample(c("Apoe1", "Apoe2", "Apoe3", "Apoe4"), 100, replace = TRUE),
  AC = rnorm(100),
  Ala = rnorm(100),
  # ... Additional variables ...
  H1 = rnorm(100)
)

categories <- data.frame(
  firID = 1:100,
  Sample.Description = sample(c("DLB", "msAD", "miAD", "HC"), 100, replace = TRUE),
  Allgr = sample(c("AD", "HC"), 100, replace = TRUE),
  Gender = sample(c("Female", "Male"), 100, replace = TRUE),
  Age = sample(20:80, 100, replace = TRUE),
  Submission.Name = sample(c("plate1", "plate2"), 100, replace = TRUE),
  Material = sample(c("serum", "sample"), 100, replace = TRUE)
)

processed_data <- dataPreparation::process_data(data, categories)

}
