% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_high_missingness.R
\name{remove_high_missingness}
\alias{remove_high_missingness}
\title{Remove columns and rows with high missingness}
\usage{
remove_high_missingness(data, col_threshold = 10, row_threshold = 10)
}
\arguments{
\item{data}{The input data frame.}

\item{col_threshold}{The threshold percentage for column-wise missing
values (default is 10`\%`).}

\item{row_threshold}{The threshold percentage for row-wise missing
values (default is 10`\%`).}
}
\value{
A list containing the original data, cleaned data, and details
of removed columns and rows.
}
\description{
This function removes columns and rows with high missing values
from the dataset.
}
