% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_na_handling.R
\name{remove_na_columns}
\alias{remove_na_columns}
\title{Remove Rows with NA in Specified Columns}
\usage{
remove_na_columns(data, columns_to_check)
}
\arguments{
\item{data}{A data frame}

\item{columns_to_check}{Vector of column names to check for NA}
}
\value{
Data frame with rows containing NA removed
}
\description{
Remove Rows with NA in Specified Columns
}
\examples{
\dontrun{
remove_na_columns(mtcars, c("mpg", "cyl"))
}

}
