% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setColAs.R
\name{setColAsFactor}
\alias{setColAsFactor}
\title{Set columns as factor}
\usage{
setColAsFactor(dataSet, cols, n_levels = 53, verbose = TRUE)
}
\arguments{
\item{dataSet}{Matrix, data.frame or data.table}

\item{cols}{a list of colnames of dataSet (or just one) to transform into factor}

\item{n_levels}{max number of levels for factor (integer, default to 53) 
set it to -1 to disable control.}

\item{verbose}{should the function log (logical, default to TRUE)}
}
\value{
dataSet (as a \code{\link{data.table}}), with specified columns set as factor or logical.
}
\description{
Set columns as factor and control number of unique element, to avoid having too large factors.
}
\details{
Control number of levels will help you to distinguish true categoricals from just characters 
that should be handle in another way.
}
\examples{
# Load messy_adult
data("messy_adult")

# we wil change education
messy_adult <- setColAsFactor(messy_adult, cols = "education")

sapply(messy_adult[, .(education)], class)
# education is now a factor
}
