% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_set.R
\name{shape_set}
\alias{shape_set}
\title{Final preparation before ML algorithm}
\usage{
shape_set(data_set, final_form = "data.table", thresh = 10, verbose = TRUE)
}
\arguments{
\item{data_set}{Matrix, data.frame or data.table}

\item{final_form}{"data.table" or "numerical_matrix" (default to data.table)}

\item{thresh}{Threshold such that  a numerical column is transformed into
  a factor whenever its number of unique modalities is smaller or equal to
\code{thresh} (numeric, default to 10)}

\item{verbose}{Should the algorithm talk? (logical, default to TRUE)}
}
\description{
Prepare a data.table by:
\itemize{
 \item transforming numeric variables into factors whenever they take less than \code{thresh} unique
   variables
 \item transforming characters using \code{\link{generate_from_character}}
 \item transforming logical into binary integers
 \item dropping constant columns
 \item Sending the data.table to \code{\link{set_as_numeric_matrix}} (when \code{final_form == "numerical_matrix"})
   will then allow you to get a numerical matrix usable by most Machine Learning Algorithms.
}
}
\section{Warning}{

All these changes will happen \strong{by reference}.
}

