% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_workflow_generator.R
\name{show_dataset_columns}
\alias{show_dataset_columns}
\title{Show Dataset Columns and Types}
\usage{
show_dataset_columns(dataset_name, envir = .GlobalEnv)
}
\arguments{
\item{dataset_name}{Name of the dataset (as string)}

\item{envir}{Environment to search (default: global environment)}
}
\value{
Invisible data frame with column information
}
\description{
Displays all columns in a dataset with their data types,
helping users select appropriate columns for analysis.
}
\examples{
\dontrun{
movies <- data.frame(title = c("The Matrix"), rating = c(8.7))
show_dataset_columns("movies")
}

}
