% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_numeric_features.R
\name{standardize_numeric}
\alias{standardize_numeric}
\title{Standardize Numeric Column (Zero Mean, Unit Variance)}
\usage{
standardize_numeric(data, col)
}
\arguments{
\item{data}{A data frame}

\item{col}{Column name (character)}
}
\value{
Data frame with standardized column
}
\description{
Standardize Numeric Column (Zero Mean, Unit Variance)
}
