% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_na_handling.R
\name{summarize_na_rows}
\alias{summarize_na_rows}
\title{Summarize NA Impact at Row Level}
\usage{
summarize_na_rows(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
Tibble with NA summary statistics
}
\description{
Summarize NA Impact at Row Level
}
\examples{
\dontrun{
summarize_na_rows(mtcars)
}

}
