% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set.R
\docType{data}
\name{tiny_messy_adult}
\alias{tiny_messy_adult}
\title{First 500 rows of \code{\link{messy_adult}}}
\format{
A data.table with 500 rows and 24 variables.
}
\usage{
data(tiny_messy_adult)
}
\description{
First 500 rows of \code{\link{messy_adult}}
}
\keyword{data}
