% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_numeric_features.R
\name{transform_log}
\alias{transform_log}
\title{Apply Logarithmic Transformation}
\usage{
transform_log(data = NULL, col, offset = 1, dataset_name = NULL)
}
\arguments{
\item{data}{A data frame (optional if dataset_name provided)}

\item{col}{Column name (character)}

\item{offset}{Offset to add before log transformation}

\item{dataset_name}{Character. Name of registered dataset to use.}
}
\value{
Data frame with log-transformed column
}
\description{
Apply Logarithmic Transformation
}
