% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_numeric_features.R
\name{transform_sqrt}
\alias{transform_sqrt}
\title{Apply Square Root Transformation}
\usage{
transform_sqrt(data, col)
}
\arguments{
\item{data}{A data frame}

\item{col}{Column name (character)}
}
\value{
Data frame with square root transformed column
}
\description{
Apply Square Root Transformation
}
