% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_batch_correction.R
\name{visualize_batch_correction}
\alias{visualize_batch_correction}
\title{Visualize Batch Correction Effects}
\usage{
visualize_batch_correction(corrected_list, visualization_type = "boxplot")
}
\arguments{
\item{corrected_list}{A list containing corrected data matrices.}

\item{visualization_type}{A character string specifying the type of visualization.
It can be either "boxplot" or "barplot".}
}
\value{
A ggplot object representing the visualization.
}
\description{
This function visualizes the effects of batch correction methods by generating boxplot or barplot.
}
\examples{
# Visualize batch correction effects using boxplot
visualize_batch_correction(corrected_list, visualization_type = "boxplot")

# Visualize batch correction effects using barplot
visualize_batch_correction(corrected_list, visualization_type = "barplot")

}
