% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregateByKey}
\alias{aggregateByKey}
\title{Automatic dataSet aggregation by key}
\usage{
aggregateByKey(dataSet, key, verbose = TRUE, ...)
}
\arguments{
\item{dataSet}{Matrix, data.frame or data.table (with only numeric, integer, factor, logical, character columns).}

\item{key}{The name of a column of dataSet according to which the set should be aggregated (character)}

\item{verbose}{Should the algorithm talk? (logical, default to TRUE)}

\item{...}{optional argument: \code{functions}:  aggregation functions for numeric columns (list of functions)
(vector of function, optional, if not set we use: c(mean, min, max, sd))}
}
\value{
A \code{\link{data.table}} with one line per key elements.
}
\description{
Automatic aggregation of a dataSet set according to a key.\cr
}
\details{
Be carefull using functions agrument, the function given should be an aggregation fuction, meaning that for multiple values it should only return one value.
}
\examples{
# Get generic dataset from R
data("adult")

# Aggregate it using aggregateByKey, in order to extract characteristics for each country
adult_aggregated <- aggregateByKey(adult, key = 'country')

# Exmple with other functions
power <- function(x){sum(x^2)}
adult_aggregated <- aggregateByKey(adult, key = 'country', functions = c(power, sqrt))

# sqrt is not an aggregation function, so it wasn't used.
}
