% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_posixct.R
\name{as.POSIXct_fast}
\alias{as.POSIXct_fast}
\title{Faster date transformation}
\usage{
as.POSIXct_fast(x, ...)
}
\arguments{
\item{x}{An object to be converted}

\item{...}{other argument to pass to  \code{\link{as.POSIXct}}}
}
\value{
as.POSIXct and as.POSIXlt return an object of the appropriate class. If tz was specified,
as.POSIXlt will give an appropriate "tzone" attribute. Date-times known to be invalid will be returned as NA.
}
\description{
Based on the trick that often dates are repeated in a column, we make date transformation
 faster by computing date transformation only on uniques.
}
\details{
The more
}
\examples{
# Work the same as as.POSIXct
as.POSIXct_fast("2018-01-01", format="\%Y-\%m-\%d")
}
