% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_numeric_features.R
\name{binning_width}
\alias{binning_width}
\title{Bin Numeric Column (Equal-Width)}
\usage{
binning_width(data, col, n_bins = 5)
}
\arguments{
\item{data}{A data frame}

\item{col}{Column name (character)}

\item{n_bins}{Number of bins}
}
\value{
Data frame with binned column
}
\description{
Bin Numeric Column (Equal-Width)
}
