% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateFromFactor.R
\name{build_target_encoding}
\alias{build_target_encoding}
\title{Build target encoding}
\usage{
build_target_encoding(dataSet, col_to_encode, target_col, functions = "mean")
}
\arguments{
\item{dataSet}{Matrix, data.frame or data.table}

\item{col_to_encode}{column to aggregate according to (character)}

\item{target_col}{column to aggregate (character)}

\item{functions}{functions of aggregation (list or character, default to "mean")}
}
\value{
A \code{\link{data.table}} with a line by unique value of \code{col_to_encode} and 
\code{len(functions) + 1} columns.
}
\description{
Target encoding is the process of replacing a categorical value with the aggregation of the  
target variable. \code{build_target_encoding} is used to compute aggregations.
}
\examples{
# Build a data set
require(data.table)
dataSet <- data.table(student = c("Marie", "Marie", "Pierre", "Louis", "Louis"), 
                      grades = c(1, 1, 2, 3, 4))

# Perform target_encoding construction
build_target_encoding(dataSet, col_to_encode = "student", target_col = "grades", 
                      functions = c("mean", "sum"))
}
