% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_numeric_features.R
\name{create_difference_features}
\alias{create_difference_features}
\title{Create Difference Features}
\usage{
create_difference_features(data, col, lag = 1)
}
\arguments{
\item{data}{A data frame}

\item{col}{Column name (character)}

\item{lag}{Lag period}
}
\value{
Data frame with difference column
}
\description{
Create Difference Features
}
