% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_numeric_features.R
\name{create_interaction_features}
\alias{create_interaction_features}
\title{Create Interaction Terms}
\usage{
create_interaction_features(data, col1, col2)
}
\arguments{
\item{data}{A data frame}

\item{col1}{First column name}

\item{col2}{Second column name}
}
\value{
Data frame with interaction term
}
\description{
Create Interaction Terms
}
