% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_numeric_features.R
\name{create_percentage_change}
\alias{create_percentage_change}
\title{Create Percentage Change Features}
\usage{
create_percentage_change(data, col, lag = 1)
}
\arguments{
\item{data}{A data frame}

\item{col}{Column name (character)}

\item{lag}{Lag period}
}
\value{
Data frame with percentage change column
}
\description{
Create Percentage Change Features
}
