% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_numeric_features.R
\name{create_polynomial_features}
\alias{create_polynomial_features}
\title{Create Polynomial Features}
\usage{
create_polynomial_features(data, col, degree = 2)
}
\arguments{
\item{data}{A data frame}

\item{col}{Column name (character)}

\item{degree}{Degree of polynomial (2 or 3)}
}
\value{
Data frame with polynomial feature columns
}
\description{
Create Polynomial Features
}
