% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_numeric_features.R
\name{create_quantile_features}
\alias{create_quantile_features}
\title{Create Quantile Features}
\usage{
create_quantile_features(data = NULL, col, dataset_name = NULL)
}
\arguments{
\item{data}{A data frame (optional if dataset_name provided)}

\item{col}{Column name (character)}

\item{dataset_name}{Character. Name of registered dataset to use.}
}
\value{
Data frame with quantile column
}
\description{
Create Quantile Features
}
