% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_timestamp_features.R
\name{extract_year_from_string}
\alias{extract_year_from_string}
\title{Extract 4-Digit Year from String Column}
\usage{
extract_year_from_string(data, string_col = "string")
}
\arguments{
\item{data}{A data frame}

\item{string_col}{String column name to extract year from}
}
\value{
Data frame with extracted year column
}
\description{
Extract 4-Digit Year from String Column
}
