% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{fastDiscretization}
\alias{fastDiscretization}
\title{Discretization}
\usage{
fastDiscretization(dataSet, cols = "auto", n_bins = 10,
  type = "equal_width", verbose = TRUE)
}
\arguments{
\item{dataSet}{Matrix, data.frame or data.table}

\item{cols}{list of numeric column(s) name(s) of dataSet to transform. To transform all 
characters, set it to "auto"}

\item{n_bins}{number of group to compute (numeric, default to 10)}

\item{type}{type of discretization ("equal_width" or "equal_freq")}

\item{verbose}{Should the algorithm talk? (Logical, default to TRUE)}
}
\value{
Same dataset edited by reference. With `cols` transformed into bins. \cr
If you don't want to edit by reference please provide set `dataSet = copy(dataSet)`.
}
\description{
Discretization of numeric variable (either equal_width or equal_fred).
}
\details{
NAs will be putted in an NA category. Using equal freq first bin will start at -Inf and last bin will end at +Inf.
}
\examples{
data(messy_adult)
head(messy_adult)

messy_adult <- fastDiscretization(messy_adult)

head(messy_adult)
}
