% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_na_handling.R
\name{flag_na_column}
\alias{flag_na_column}
\title{Flag NA Values in Column}
\usage{
flag_na_column(data, col)
}
\arguments{
\item{data}{A data frame}

\item{col}{Column name (character)}
}
\value{
Data frame with new NA flag column
}
\description{
Flag NA Values in Column
}
\examples{
\dontrun{
flag_na_column(mtcars, "mpg")
}

}
