% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_numeric_features.R
\name{flag_outliers_iqr}
\alias{flag_outliers_iqr}
\title{Flag Outliers (IQR Method)}
\usage{
flag_outliers_iqr(data, col, multiplier = 1.5)
}
\arguments{
\item{data}{A data frame}

\item{col}{Column name (character)}

\item{multiplier}{IQR multiplier}
}
\value{
Data frame with outlier flag
}
\description{
Flag Outliers (IQR Method)
}
