% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateFromCharacter.R
\name{generateFromCharacter}
\alias{generateFromCharacter}
\title{Recode character}
\usage{
generateFromCharacter(dataSet, cols, verbose = TRUE, drop_cols = FALSE, ...)
}
\arguments{
\item{dataSet}{Matrix, data.frame or data.table}

\item{cols}{a list of character colnames of dataSet (or just one) to transform into dates. \cr
To transform all characters, set it to "auto"}

\item{verbose}{should the function log (logical, default to TRUE)}

\item{drop_cols}{should \code{cols} be dropped after generation (logical, default to FALSE)}

\item{...}{Other arguments such as \code{name_separator} to separate words in new coluimns names
(character, default to ".")}
}
\description{
Recode character into 3 new columns: \cr
\itemize{
\item was the value not NA, "NA", ""
\item how often this value occures
\item the order of the value (ex: M/F => 2/1 because F comes before M in alphabet)
}
}
\examples{
# Load data set
data(messy_adult)
messy_adult <- unFactor(messy_adult, verbose = FALSE) # un factor ugly factors

# transform column "mail"
messy_adult <- generateFromCharacter(messy_adult, cols = "mail")
head(messy_adult)

# To transform all characters columns:
messy_adult <- generateFromCharacter(messy_adult, cols = "auto")
}
