% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSet.R
\name{prepareSet}
\alias{prepareSet}
\title{Preparation pipeline}
\usage{
prepareSet(dataSet, finalForm = "data.table", verbose = TRUE, ...)
}
\arguments{
\item{dataSet}{Matrix, data.frame or data.table}

\item{finalForm}{"data.table" or "numerical_matrix" (default to data.table)}

\item{verbose}{Should the algorithm talk? (logical, default to TRUE)}

\item{...}{additional parameters to thune pipeline (see details)}
}
\value{
A data.table or a numerical matrix (according to finalForm)  and
}
\description{
Full pipeline for preparing your dataSet set \cr
It will perform the following steps: \cr
- Correct set: id dates and numerics that are hiden in string \cr
- Transform set: compute differences between every date, if `key` is provided, 
will perform aggregate according to this key \cr
- Filter set: filter constant, in double or bijection variables. If `digits` is provided, 
will round numerics \cr
- Handle NA: will perform \code{\link{fastHandleNa}}) \cr
- Shape set: will put the result in asked shape (`finalForm`) with acceptable columns format.
}
\details{
Aditional arguments are available to thune pipeline: 
\itemize{
  \item \code{key} name of a column of dataSet according to which dataSet should be aggregated (character)
  \item \code{analysisDate} A date at which the dataSet should be aggregated 
(differences between every date and analysisDate will be computed) (Date)
  \item \code{digits} The number of digits after comma (optional, numeric, if set will perform \code{\link{fastRound}})
  \item \code{dateFormats} List of format of Dates in dataSet (list of characters)
  \item \code{name_separator} string to separate parts of new column names (string)
}
}
\examples{
# Load ugly set
data(messy_adult)

# Have a look to set
head(messy_adult)

# Compute full pipeline
clean_adult <- prepareSet(messy_adult)

# With a reference date
adult_agg <- prepareSet(messy_adult, analysisDate = as.Date("2017-01-01"))

# Add aggregation by country
adult_agg <- prepareSet(messy_adult, analysisDate = as.Date("2017-01-01"), key = "country")

# With some new aggregation functions
power <- function(x){x^2}
adult_agg <- prepareSet(messy_adult, analysisDate = as.Date("2017-01-01"), key = "country", 
                        functions = c("min", "max", "mean", "power"))
}
