% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_numeric_features.R
\name{transform_log}
\alias{transform_log}
\title{Apply Logarithmic Transformation}
\usage{
transform_log(data, col, offset = 1)
}
\arguments{
\item{data}{A data frame}

\item{col}{Column name (character)}

\item{offset}{Offset to add before log transformation}
}
\value{
Data frame with log-transformed column
}
\description{
Apply Logarithmic Transformation
}
